Provide means to change a line based file.
LineInterface has a concrete implementation NameValueLine, based on
AbstractLine, and which allows a line in a file to be added or modified
and which signals when a change to the line has occurred.

LineBasedFileInterface has a concrete implementation LineBasedFile to
represent a file and which reads and writes LineInterface lines,
signalling when a change to the file has occurred.